Increase amount of memory available for tests in self-hosted runners
Our self-hosted runners have a lot of memory allocated to the
/var/lib/docker which is tmpfs mounted storage for docker engine.

This is done in order to speed up immensely any docker-related
operations - such as building and creating images and running and
deploying kubernetes instances.

The memory allocated was at the 85% of capacity ~ 52 GB. However
it seems that with our setup when we run up to two full K8S clusters,
we are peaking at ~40GB . We can safely allocate more memory for
tests and other operations - which might speed up the speed
of tests.